% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arx_boot.R
\name{arx_boot}
\alias{arx_boot}
\title{Run the bootstrap algorithm on simulated ARX data to get an estimate for the covariance between variables and parameters.}
\usage{
arx_boot(
  boot_reps = 100,
  sim_length = 1000,
  params = list(alpha = 0.5, beta = c(0.2, 0.8), eta = c(0.3, 0.4, 0.5))
)
}
\arguments{
\item{boot_reps}{An integer indicating the number of bootstrap replications to do}

\item{params}{A named list indicating the true parameters to use for the simulated data}
}
\value{
A list with each entry being one bootstrap replicate of the simulated data
}
\description{
Run the bootstrap algorithm on simulated ARX data to get an estimate for the covariance between variables and parameters.
}
\examples{
arx_boot(500)
}
