% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arx_mediation_boot.R
\name{arx_mediation_boot}
\alias{arx_mediation_boot}
\title{Perform bootstrapping on the residuals of a time series process and return the specified number of bootstrap replicates of the time series}
\usage{
arx_mediation_boot(data, p, q = c(0, 0, 0), R)
}
\arguments{
\item{data}{A data frame containing vectors of time series data}

\item{p}{A vector of values for the degree of the ar model for the X,M,Y}

\item{q}{A vector of values for the degree of the ma model for the X,M,Y, defaults to 0}

\item{R}{An integer for the number of bootstrap replicates used}
}
\value{
A named list which includes an estimate of the time series, and the bootstrap results
}
\description{
Perform bootstrapping on the residuals of a time series process and return the specified number of bootstrap replicates of the time series
}
\examples{
data <- arx_simulation(100)
data <- data.frame(data$X, data$M, data$Y)
}
