% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arx_plotting.R
\name{arx_plotting}
\alias{arx_plotting}
\title{Plot the results of \code{arx_summary}, giving a violin plot for the parameters}
\usage{
arx_plotting(results)
}
\arguments{
\item{results}{A named list of lists, each one being a property of the data}
}
\value{
Three plots, one for each variable
}
\description{
Plot the results of \code{arx_summary}, giving a violin plot for the parameters
}
\examples{
boots <- arx_boot(500, params = list(alpha = 0.2, beta = c(0.1,0.5), eta = c(0.2,0.3,0.6)))
boots2 <- arx_summary(boots, params = list(alpha = 0.2, beta = c(0.1,0.5), eta = c(0.2,0.3,0.6)))
arx_plotting(boots2)
}
