% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arx_summary.R
\name{arx_summary}
\alias{arx_summary}
\title{Summarize the results of \code{arx_boot} to get estimates of the covariances}
\usage{
arx_summary(
  boots,
  params = list(alpha = 0.5, beta = c(0.2, 0.8), eta = c(0.3, 0.4, 0.5))
)
}
\arguments{
\item{boots}{A list of data frames, each one being a bootstrap replicate of the data}
}
\value{
Named list
}
\description{
Summarize the results of \code{arx_boot} to get estimates of the covariances
}
\examples{
boots <- arx_boot(500)
arx_summary(boots)
}
