% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_practice.R
\name{lags}
\alias{lags}
\title{Perform regression on lagged data}
\usage{
lags(formula, df)
}
\arguments{
\item{df}{A data frame an x and y variable}

\item{n}{An integer indicating what degree of lag to use}
}
\value{
A linear model of y on x and n lagged x terms
}
\description{
Perform regression on lagged data
}
\examples{
x <- rnorm(100)
y <- runif(100)
df <- data.frame(x,y)
lags(df,3)
}
