% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediation_spline.R, R/mediation_spline_res.R
\name{mediation_spline}
\alias{mediation_spline}
\title{Perform a spline mediation on variables X, M, and Y}
\usage{
mediation_spline(df)

mediation_spline(df)
}
\arguments{
\item{df}{A data frame with columns X, M, and Y}
}
\value{
Values for the total effect, direct effect, the indirect effect calculated in two different ways, and standard error for each of these values.
Also returns a covariance matrix for those values.

Values for the total effect, direct effect, the indirect effect calculated in two different ways, and standard error for each of these values.
Also returns a covariance matrix for those values.
}
\description{
Perform a spline mediation on variables X, M, and Y

Perform a spline mediation on variables X, M, and Y
}
\examples{
mediation_spline(df)
mediation_spline(df)
}
