% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_rf.R
\name{prediction_rf}
\alias{prediction_rf}
\title{Use a prediction based method along with random forest models to find the direct and indirect effects of a mediation system}
\usage{
prediction_rf(data, p = 10)
}
\arguments{
\item{boot_reps}{An integer indicating the number of bootstrap replications to do}

\item{params}{A named list indicating the true parameters to use for the simulated data}
}
\value{
A list with each entry being one bootstrap replicate of the simulated data
}
\description{
Use a prediction based method along with random forest models to find the direct and indirect effects of a mediation system
}
\examples{
boots <- arx_boot(500)
prediction_rf(boots[[1]])

or
sims <- arx_simulation(100)
dat <- data.frame(sims$X, sims$M, sims$Y)
prediction_rf(dat)

}
