% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_boot_ts.R
\name{run_boot_ts}
\alias{run_boot_ts}
\title{Perform bootstrapping on the residuals of a time series process}
\usage{
run_boot_ts(x, R)
}
\arguments{
\item{x}{residuals from a time series process}

\item{R}{number of bootstrap replicates}
}
\value{
A data frame of bootstrap replicates of the residuals
}
\description{
Perform bootstrapping on the residuals of a time series process
}
\examples{
x <- residuals(ts_object)
run_boot_ts(x, R)
}
