% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_mediation.R
\name{simple_mediation}
\alias{simple_mediation}
\title{Performs simple mediation}
\usage{
simple_mediation(df)
}
\arguments{
\item{df}{A data frame with columns X, M, and Y}
}
\value{
Values for the total effect, direct effect, the indirect effect calculated in two different ways, and standard error for each of these values.
}
\description{
Performs simple mediation
}
\examples{
x <- c(1,2,3,4,5,6,7)
m <- c(1,2.3,4.2,4.9,5.6,6,6.7)
y <- c(3.4,5.4,6.5,7.9,8,10,11)
df <- data.frame(x,m,y)
simple_mediation(df)
}
