% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediation_spline.R, R/mediation_spline_res.R
\name{single_spline}
\alias{single_spline}
\title{A helper function to run a single iteration of spline mediation, which will then be used to perform bootstrapping on}
\usage{
single_spline(df, samps, knots = list(seq(min(x), max(x), length = 5, k)))

single_spline(df, samps, knots = list(seq(min(x), max(x), length = 5, k)))
}
\description{
A helper function to run a single iteration of spline mediation, which will then be used to perform bootstrapping on

A helper function to run a single iteration of spline mediation, which will then be used to perform bootstrapping on
}
