% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_mediation_boot.R
\name{ts_mediation_boot}
\alias{ts_mediation_boot}
\title{Perform bootstrapping on the residuals of a time series process and return the specified number of bootstrap replicates of the time series. Currently does it for 3 time series, but does not for causal relationships between them.}
\usage{
ts_mediation_boot(data, p, q = c(0, 0, 0), R)
}
\arguments{
\item{data}{A data frame containing vectors of time series data}

\item{p}{A vector of values for the degree of the ar model for the X,M,Y}

\item{q}{A vector of values for the degree of the ma model for the X,M,Y, defaults to 0}

\item{R}{An integer for the number of bootstrap replicates used}
}
\value{
A named list which includes an estimate of the time series, and the bootstrap results
}
\description{
Perform bootstrapping on the residuals of a time series process and return the specified number of bootstrap replicates of the time series. Currently does it for 3 time series, but does not for causal relationships between them.
}
\examples{
x <- rnorm(100)
m <- rnorm(100)
y <- rnorm(100)
data <- data.frame(x,m,y)
ts_mediation_boot(data, p = c(1,1,1), R = 50)
}
