% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_resid_boot.R
\name{ts_resid_boot}
\alias{ts_resid_boot}
\title{Perform bootstrapping on the residuals of a time series process and return the specified number of bootstrap replicates of the time series}
\usage{
ts_resid_boot(x, n, R)
}
\arguments{
\item{x}{A vectors of time series data}

\item{n}{An integer indicating the degree of AR model given}

\item{R}{An integer for the number of bootstrap replicates used}
}
\value{
A named list which includes an estimate of the time series, and the bootstrap results
}
\description{
Perform bootstrapping on the residuals of a time series process and return the specified number of bootstrap replicates of the time series
}
\examples{
x <- rnorm(100)
ts_simple_resid_boot(x, 2, 50)
}
