% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_simple_resid_boot.R
\name{ts_simple_resid_boot}
\alias{ts_simple_resid_boot}
\title{Perform bootstrapping on the residuals of a time series process}
\usage{
ts_simple_resid_boot(df)
}
\arguments{
\item{df}{A data frame with a value changing over time}
}
\value{
An estimate of the time series, and the bootstrap results
}
\description{
Perform bootstrapping on the residuals of a time series process
}
\examples{
x <- rnorm(100)
ts_simple_resid_boot(x)
}
