% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_mediation.R
\name{run_mediation}
\alias{run_mediation}
\title{Run a simple linear Mediation}
\usage{
run_mediation(df, samps)
}
\arguments{
\item{df}{a vector of observations}

\item{samps}{The sample indices, as required by boot::boot()}
}
\value{
An estimate of the total, direct, and two estimates of the indirect effect in that order
}
\description{
Run a simple linear Mediation
}
\examples{
x <- c(1,2,3,4,5,6,7)
m <- c(1,2.3,4.2,4.9,5.6,6,6.7)
y <- c(3.4,5.4,6.5,7.9,8,10,11)
df <- data.frame(x,m,y)
run_mediation(df)
}
